draft: change & apply sidebar to be only navigation
