--- src/components/organisms/Sidebar/index.test.tsx
@@ -1,30 +1,0 @@
-import { cleanup, render, screen } from '@testing-library/react';
-import { RouterProvider, createMemoryRouter } from 'react-router-dom';
-import { beforeEach, describe, expect, it } from 'vitest';
-
-import { rootNavNodes } from '^/constants/nav-node';
-
-import { Sidebar } from '.';
-
-describe('Sidebar', () => {
-  const routes = [
-    {
-      path: '/',
-      element: <Sidebar rootNavNodes={rootNavNodes} />,
-    },
-  ];
-
-  const router = createMemoryRouter(routes, {
-    initialEntries: ['/'],
-    initialIndex: 0,
-  });
-
-  beforeEach(() => {
-    cleanup();
-  });
-
-  it('should show title on init', () => {
-    render(<RouterProvider router={router} />);
-    expect(screen.getByText(/YSOShmupRecords/i)).not.toBeNull();
-  });
-});
