--- src/components/organisms/Sidebar/index.tsx
@@ -1,56 +1,0 @@
-import { useState } from 'react';
-import styled from 'styled-components';
-
-import { NavigationForest } from '^/components/molecules/NavigationForest';
-import { SidebarFooter } from '^/components/molecules/SidebarFooter';
-import { SidebarHeader } from '^/components/molecules/SidebarHeader';
-import { TitleWithAvatar } from '^/components/molecules/TitleWithAvatar';
-import { NavNodeInfo } from '^/types';
-
-const Root = styled.div`
-  height: 100%;
-
-  display: grid;
-  grid-template-rows: auto 1fr auto;
-
-  background: var(--primary-color);
-`;
-
-const CenterWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-`;
-
-interface Props {
-  rootNavNodes: NavNodeInfo[];
-}
-
-export function Sidebar({ rootNavNodes }: Props) {
-  const [isNavigationOpen, setIsNavigationOpen] = useState<boolean>(false);
-
-  const renderCenter = isNavigationOpen ? (
-    <NavigationForest
-      onClickNavigationNode={() => {
-        setIsNavigationOpen(false);
-      }}
-      rootNavNodes={rootNavNodes}
-    />
-  ) : (
-    <TitleWithAvatar />
-  );
-
-  return (
-    <Root>
-      <SidebarHeader
-        isNavigationOpen={isNavigationOpen}
-        onClickOpenOrCloseNavigationButton={() => {
-          setIsNavigationOpen(!isNavigationOpen);
-        }}
-      />
-      <CenterWrapper>{renderCenter}</CenterWrapper>
-      <SidebarFooter />
-    </Root>
-  );
-}
